Update fix_nh_constant_pH.h
An intermediate file has been added to the fix_constant_pH.cpp and fix_nh_constant_pH.cpp to read and write the molids before/after the minimization whenever the protonable molids change. 
A bug was solved that causes the t_target to go infinite.
And many more bug fixes.